+++ src/test/java/lotto/WinningTest.java
@@ -54,10 +54,9 @@ public class WinningTest {
     @Test
     void yieldTest() {
         //given
-        int buyMoney = 14000;
+        Long buyMoney = 14000L;
         int[] match = new int[]{0, 1, 2, 3, 4, 5, 6};
-        int expectMatch = 0 * 0 + 1 * 0 + 2 * 0
-                + 3 * THREE_MATCH_PRICE + 4 * FOUR_MATCH_PRICE + 5 * FIVE_MATCH_PRICE + 6 * SIX_MATCH_PRICE;
+        Long expectMatch = 0 * 0 + 1 * 0 + 2 * 0 + 3 * THREE_MATCH_PRICE + 4 * FOUR_MATCH_PRICE + 5 * FIVE_MATCH_PRICE + 6 * SIX_MATCH_PRICE;
         double expect =  round(expectMatch / buyMoney * 100) / 100.0;
         Winning winning = new Winning(winningNumbers, match);
